change data type of rate and rateOriginal to float

+++ AxonsERP.Entities/DataTransferObjects/TaxRateControlDto.cs
@@ -9,11 +9,11 @@ namespace AxonsERP.Entities.DataTransferObjects
         [Required]
         public DateTime effectiveDate { get; set; }
 
-        public int rate { get; set; }
+        public float rate { get; set; }
         public string? owner { get; set; }
         public DateTime createDate { get; set; }
         public DateTime lastUpdateDate { get; set; }
         public string? function { get; set; }
-        public int rateOriginal { get; set; }
+        public float rateOriginal { get; set; }
     }
 }
